Resolves #234 - Set default git protocol to https
The recommended protocol for communicating with github is https.  This
change sets the default protocol for cloning the repository to https.

Add command line options '--use-git' and '--use-ssh' to install.sh.
This will allow users to specify which protocol to use when cloning
the repository.

install.sh will now print the `git clone` command to stdout.